Sidebar: Show sidebar item as selected when query term present
Currently, if you're on a route like https://wordpress.com/pages and you
use the search feature to search through all of your site pages, the
Pages sidebar item no longer appears selected despite still being on a
pages route. This updates the logic used to determine whether a sidebar
item should appear selected. It now returns truthy if the path of the
item is a prefix of the current path plus a search query.

Fixes: #18153